feat: add metadata for pages and sections
